[refactor] findByPlaceIds -> findByPlaceIdIn으로 변경
- videoRepository에서 제목과같이 변경 후, PlaceService에 적용하였습니다.

관련 이슈: #7
