+++ src/main/java/team7/inplace/place/domain/PlaceCloseTime.java
@@ -1,0 +1,2 @@
+package team7.inplace.place.domain;public class PlaceCloseTime {
+}
