+++ src/main/java/team7/inplace/video/persistence/VideoRepository.java
@@ -2,8 +2,6 @@ package team7.inplace.video.persistence;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import team7.inplace.place.domain.Place;
 import team7.inplace.video.domain.Video;
 
@@ -11,11 +9,9 @@ public interface VideoRepository extends JpaRepository<Video, Long> {
 
     List<Video> findVideosByInfluencerIdIn(List<Long> influencerIds);
 
-    @Query("SELECT v FROM Video v WHERE v.place.id IN :placeIds")
-    List<Video> findByPlaceIds(@Param("placeIds") List<Long> placeIds);
-
     List<Video> findAllByOrderByIdDesc();
 
     Video findTopByPlaceOrderByIdDesc(Place place);
 
+    List<Video> findByPlaceIdIn(List<Long> placeIds);
 }
